got starter to work. added icons and favicons
